Renamed folder PollingStation to polling_station and moved form pages to page folder
