--- frontend/app/module/data_entry/PollingStation/index.ts
@@ -1,5 +1,0 @@
-export * from "./CandidatesVotesPage";
-export * from "./DifferencesPage";
-export * from "./PollingStationLayout";
-export * from "./RecountedPage";
-export * from "./VotersAndVotesPage";
