+++ frontend/app/module/data_entry/index.ts
@@ -1,3 +1,3 @@
 export * from "./DataEntryLayout";
 export * from "./page";
-export * from "./PollingStation";
+export * from "./polling_station";
