+++ frontend/app/module/data_entry/page/CandidatesVotesPage.tsx
