+++ frontend/app/module/data_entry/page/DifferencesPage.tsx
