+++ frontend/app/module/data_entry/page/VotersAndVotesPage.tsx
