+++ frontend/app/module/data_entry/page/index.ts
@@ -1,1 +1,5 @@
+export * from "./CandidatesVotesPage";
 export * from "./DataEntryHomePage";
+export * from "./DifferencesPage";
+export * from "./RecountedPage";
+export * from "./VotersAndVotesPage";
