+++ frontend/app/module/data_entry/polling_station/AbortDataEntryControl.test.tsx
@@ -5,6 +5,7 @@ import { http, HttpResponse } from "msw";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 
 import { VotersAndVotesForm } from "app/component/form/data_entry/voters_and_votes/VotersAndVotesForm";
+import { AbortDataEntryControl } from "app/module/data_entry";
 import { overrideOnce, render, screen, server } from "app/test/unit";
 import { emptyDataEntryRequest } from "app/test/unit/form";
 
@@ -16,8 +17,6 @@ import {
   SaveDataEntryResponse,
 } from "@kiesraad/api";
 import { electionDetailsMockResponse, electionMockData, pollingStationMockData } from "@kiesraad/api-mocks";
-
-import { AbortDataEntryControl } from "./AbortDataEntryControl";
 
 const mockNavigate = vi.fn();
