+++ frontend/app/module/data_entry/polling_station/AbortDataEntryModal.tsx
